% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBGP.R
\name{as_func_call}
\alias{as_func_call}
\title{Rule action for constructing a function call.}
\usage{
as_func_call(vs)
}
\arguments{
\item{vs}{The list of values, the first of which will be treated as
function using \code{match.fun()}, and the rest will be treated
as arguments.}
}
\value{
an R function call expression.
}
\description{
Rule action for constructing a function call.
}
