% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBGP.R
\name{cal_min_heights}
\alias{cal_min_heights}
\title{Helper function to calculate the minimum heights of each
non-terminal and grammar rule. A terminal has height 1, and each
level of non-terminal adds 1 to the height.}
\usage{
cal_min_heights(rules)
}
\arguments{
\item{rules}{The named list of grammar rules with names as returned
by \code{normlize_and_name_rules()}.}
}
\value{
A named list of:

nt_min_heights: named list of minimum heights of non-terminals,
using the non-terminal name as name

rules_min_heights: named list of minimum heights of the rules,
using the rule id as name
}
\description{
Helper function to calculate the minimum heights of each
non-terminal and grammar rule. A terminal has height 1, and each
level of non-terminal adds 1 to the height.
}
