% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBGP.R
\name{chr_crossover_func}
\alias{chr_crossover_func}
\title{A simple chromosome crossover function, that can serve as reference
for customization. This function will first select a node from the
first chromosome, and then select a node of the same non-terminal
in the second chromosome, and then crossover them. If no nodes of
the same non-terminal in the second chromosome can be found, the
function will retry by selecting another node from the first
chromosome, up to a maximum number of trials. If there is no
success after a maximum number of trials, the first chromosome will
simply be returned. If the first chromosome has no nodes (which
should not happen), then the second chromosome will be returned. In
the simplest case, the crossover simply use the selected node from
the second chromosome to replace the selected node in the first
chromosome, but custom node crossover function specific to the
non-terminal could be provided, see \code{nt_crossovers} for
details.}
\usage{
chr_crossover_func(
  chr1,
  chr2,
  n.tries = 10,
  is_wanted1 = always_true,
  is_wanted2 = always_true,
  height_prob_func1 = NULL,
  height_prob_func2 = height_prob_func1,
  nt_crossovers = list()
)
}
\arguments{
\item{chr1}{The first chromosome.}

\item{chr2}{The second chromosome.}

\item{n.tries}{The maximum number of tries before giving up and
return the first chromosome.}

\item{is_wanted1}{Optional predicate function whether to consider
nodes of chr1.}

\item{is_wanted2}{Optional predicate function whether to consider
nodes of chr2.}

\item{height_prob_func1}{Optional, default NULL. If non-NULL,
should be a function(heights, nodes) that should return the
vector of probability of choosing the nodes of chr1. If NULL,
there is no bias in the nodes.}

\item{height_prob_func2}{Optional, similar to height_prob_func1,
but for biasing the nodes of chr2.}

\item{nt_crossovers}{Optional, default empty list. This can be a
named list to maps non-terminal name to custom
function(node_chr1, node_chr2) that returns a new node from
selected nodes from chr1 and chr2, to replace the old node from
chr1. If the selected non-terminal has no entry in
\code{nt_crossovers}, then simply the node selected from chr2 is
used as the new node.}
}
\value{
A possibly crossovered chromsome.
}
\description{
A simple chromosome crossover function, that can serve as reference
for customization. This function will first select a node from the
first chromosome, and then select a node of the same non-terminal
in the second chromosome, and then crossover them. If no nodes of
the same non-terminal in the second chromosome can be found, the
function will retry by selecting another node from the first
chromosome, up to a maximum number of trials. If there is no
success after a maximum number of trials, the first chromosome will
simply be returned. If the first chromosome has no nodes (which
should not happen), then the second chromosome will be returned. In
the simplest case, the crossover simply use the selected node from
the second chromosome to replace the selected node in the first
chromosome, but custom node crossover function specific to the
non-terminal could be provided, see \code{nt_crossovers} for
details.
}
