% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBGP.R
\name{chr_mutation_func}
\alias{chr_mutation_func}
\title{A simple chromosome mutation function, that can serve as reference
for customization. This function will select a node from the
chromosome, then either the default node_mutator is applied or a
custom non-terminal specific mutator is applied if there is
one. Note that the mutation is applied only to a node corresponding
to non-terminal. If mutation of terminal is desired, then the
non-terminal specific custom mutator could be used to copy the node
but mutator one of the terminal, as appripriate.}
\usage{
chr_mutation_func(
  chr,
  default_node_mutator,
  is_wanted = always_true,
  height_prob_func = NULL,
  nt_mutators = list()
)
}
\arguments{
\item{chr}{The chromosome to mutate.}

\item{default_node_mutator}{The function(node) that mutate a
selected node, and the old node in \code{chr} will be replaced
with the mutated node.}

\item{is_wanted}{Optional predicate function whether to consider
nodes of chr.}

\item{height_prob_func}{Optional, default NULL. If non-NULL, should
be a function(heights, nodes) that should return the vector of
probability of choosing the nodes of chr. If NULL, there is no
bias in the nodes.}

\item{nt_mutators}{Optional, default empty list. This can be a
named list to maps non-terminal name to custom function(node)
that returns a new node from selected node from chr, to replace
the old node from chr. If the selected non-terminal has no entry
in \code{nt_mutators}, then simply the
\code{default_node_mutator} is applied to get the new node.}
}
\value{
A possibly mutated chromosome.
}
\description{
A simple chromosome mutation function, that can serve as reference
for customization. This function will select a node from the
chromosome, then either the default node_mutator is applied or a
custom non-terminal specific mutator is applied if there is
one. Note that the mutation is applied only to a node corresponding
to non-terminal. If mutation of terminal is desired, then the
non-terminal specific custom mutator could be used to copy the node
but mutator one of the terminal, as appripriate.
}
