% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBGP.R
\name{convert_to_phenotype}
\alias{convert_to_phenotype}
\title{To convert from genotype to phenotype}
\usage{
convert_to_phenotype(x, G)
}
\arguments{
\item{x}{The chromosome (tree of node) or any of its sub-nodes.}

\item{G}{The grammar as returned by \code{grammar()}, so as to
lookup the rule actions.}
}
\value{
The converted phenotype, which depends on the rule actions.
}
\description{
To convert from genotype to phenotype
}
