% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBGP.R
\name{create_individual}
\alias{create_individual}
\title{Convenience function to create an individual, which is a chromosome
together with fitness.}
\usage{
create_individual(chr, fitness)
}
\arguments{
\item{chr}{The chromosome.}

\item{fitness}{The fintess value.}
}
\value{
An object of class "individual".
}
\description{
Convenience function to create an individual, which is a chromosome
together with fitness.
}
