% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBGP.R
\name{create_node}
\alias{create_node}
\title{To create an internal node corresponding to a non-terminal. Note
that terminal can be directly represented.}
\usage{
create_node(nt, rule_name, children)
}
\arguments{
\item{nt}{The non-terminal as a character string.}

\item{rule_name}{The name of the rule generating the children.}

\item{children}{The list of children objects, which could be a node
(for subtree corresponding to non-terminal) or other values
(corresponding to terminals).}
}
\value{
An object of class "node", which is a tree.
}
\description{
To create an internal node corresponding to a non-terminal. Note
that terminal can be directly represented.
}
