% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBGP.R
\name{first}
\alias{first}
\title{Rule action to get the first value.}
\usage{
first(vs)
}
\arguments{
\item{vs}{The list of value, where there should only be one, and
will be returned as the constructed value.}
}
\value{
The first value.
}
\description{
Rule action to get the first value.
}
