% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBGP.R
\name{first_as_symbol}
\alias{first_as_symbol}
\title{Rule action to get leaf as symbol.}
\usage{
first_as_symbol(vs)
}
\arguments{
\item{vs}{The list of value, where there should only be one, and
will be coerced to symbol and returned as the constructed value.}
}
\value{
The first value as symbol.
}
\description{
Rule action to get leaf as symbol.
}
