% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBGP.R
\name{gen_term}
\alias{gen_term}
\title{To construct a terminal with generated value using a function
instead of a simple constant value.}
\usage{
gen_term(func)
}
\arguments{
\item{func}{The function of no arguments, to generate a value for
the terminal each time the body of a rule is instantiated.}
}
\value{
An object of class c("generated_terminal", "terminal").
}
\description{
To construct a terminal with generated value using a function
instead of a simple constant value.
}
