% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBGP.R
\name{generate_chromosome.character}
\alias{generate_chromosome.character}
\title{For character string, either generate as non-terminal, or a
terminal.}
\usage{
\method{generate_chromosome}{character}(x, G, max_height)
}
\arguments{
\item{x}{The object to create from, which could be a non-terminal,
terminal, a grammar, a rule.}

\item{G}{The grammar as returned by \code{grammar()}.}

\item{max_height}{The allowed maximum height. The generation will
try to avoid generating a chromsome with height exceeding this. A
terminal has height 1, and every layer of non-terminal adds 1 to
the height.}
}
\description{
For character string, either generate as non-terminal, or a
terminal.
}
