% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBGP.R
\name{generate_init_chrs}
\alias{generate_init_chrs}
\title{To generate a list of n chromosomes (note that it is not evaluated
yet).}
\usage{
generate_init_chrs(n, G, max_height, start_nt = G$start)
}
\arguments{
\item{n}{The number of chromosomes to generate.}

\item{G}{The grammar as returned by \code{grammar()}.}

\item{max_height}{The maximum allowed height in generation. A leaf
has height 1, each level of non-terminal adds 1 to the height.}

\item{start_nt}{Optional, the non-terminal to generate. Default is
the starting symbol of the grammar \code{G}.}
}
\value{
A list of \code{n} chromosomes.
}
\description{
To generate a list of n chromosomes (note that it is not evaluated
yet).
}
