% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBGP.R
\name{get_chr_nodes}
\alias{get_chr_nodes}
\title{To get all the nodes of a chromosome, for selection purpose.}
\usage{
get_chr_nodes(chr, is_wanted = always_true)
}
\arguments{
\item{chr}{The chromosome, which should be a node.}

\item{is_wanted}{Optional predicate function the returns TRUE if a
node is wanted.}
}
\value{
a named list with:

nodes: the list of nodes

heights: a vector of the heights of the nodes corresponding to
the nodes. This will be useful to adjust the selection
probability by height.
}
\description{
To get all the nodes of a chromosome, for selection purpose.
}
