% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBGP.R
\name{get_height_prob_func}
\alias{get_height_prob_func}
\title{An example of generating height bias function. The root has a
separate weight as bias, and each nodes of height 2 and onward gets
exponential weight. Lastly the weights are normalized to get
probability bias.}
\usage{
get_height_prob_func(subtree_weight, root_weight = 1)
}
\arguments{
\item{subtree_weight}{The exponential multiplier of nodes with
height 2 or more, i.e. (subtree_weight)^(height - 1).}

\item{root_weight}{The weight for root node, i.e. height of 1.}
}
\value{
A function that can be used as \code{height_prob_func1} and
\code{height_prob_func2} in \code{chr_crossover_func}.
}
\description{
An example of generating height bias function. The root has a
separate weight as bias, and each nodes of height 2 and onward gets
exponential weight. Lastly the weights are normalized to get
probability bias.
}
