% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBGP.R
\name{get_re_gen_node_mutator}
\alias{get_re_gen_node_mutator}
\title{An example convenience function to get a default node mutator that
re-generates the node for the non-terminal, that can be used as
\code{default_node_mutator} in \code{chr_mutation_func}.}
\usage{
get_re_gen_node_mutator(G, max_height = 5)
}
\arguments{
\item{G}{The grammar as returned by \code{grammar()}.}

\item{max_height}{The allowed maximum height in the re-generation
of the node.}
}
\value{
A function(node) that can be used as
\code{default_node_mutator} in \code{chr_mutation_func}.
}
\description{
An example convenience function to get a default node mutator that
re-generates the node for the non-terminal, that can be used as
\code{default_node_mutator} in \code{chr_mutation_func}.
}
