% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBGP.R
\name{is_non_terminal}
\alias{is_non_terminal}
\title{Whether an object is a non-terminal in a grammar.}
\usage{
is_non_terminal(x, rules)
}
\arguments{
\item{x}{The object to test.}

\item{rules}{The rules as in \code{grammar()}.}
}
\value{
TRUE if \code{x} is a non-terminal in the rules.
}
\description{
Whether an object is a non-terminal in a grammar.
}
