% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBGP.R
\name{normalize_and_name_rules}
\alias{normalize_and_name_rules}
\title{Helper function to normalize the rules.}
\usage{
normalize_and_name_rules(rules)
}
\arguments{
\item{rules}{The named list of rules as accepted by
\code{grammar()}.}
}
\value{
Normalized named list of rules, such that each rule has a
generated name if it does not already have a name, and each is a
list of rule as constructed by \code{rule()}, even though a
non-terminal may have only one rule.
}
\description{
Helper function to normalize the rules.
}
