% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBGP.R
\name{progress_reporter}
\alias{progress_reporter}
\title{Convenience function to generate a reporting function, which is a
function that gets the same inputs as stopping condition function
(pop, n_gen_ended, cur_best_fitness, n_gen_of_best_fitness,
start_time) and is called after each generation, and it should
report the progress as appropriate.}
\usage{
progress_reporter(report_every_n_gen = 1)
}
\arguments{
\item{report_every_n_gen}{Optional, default 1. Determines the
frequency to report progress. After the end of the first
generation, will report again after \code{report_every_n_gen}
generations. I.e. for the default value of 1, will report at the
end of every generation.}
}
\value{
A reporting function.
}
\description{
Convenience function to generate a reporting function, which is a
function that gets the same inputs as stopping condition function
(pop, n_gen_ended, cur_best_fitness, n_gen_of_best_fitness,
start_time) and is called after each generation, and it should
report the progress as appropriate.
}
