% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBGP.R
\name{replace_node}
\alias{replace_node}
\title{To replace old node with new node in chromosome.}
\usage{
replace_node(chr, old_node, new_node)
}
\arguments{
\item{chr}{The chromosome to copy and modify.}

\item{old_node}{The old node in \code{chr} to replace.}

\item{new_node}{The new node to replace \code{old_node} in
\code{chr}.}
}
\value{
A copied and modified chromosome with \code{old_node}
replaced with \code{new_node}.
}
\description{
To replace old node with new node in chromosome.
}
