% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBGP.R
\name{rule}
\alias{rule}
\title{To construct one rule for a non-terminal.}
\usage{
rule(..., name = NULL, action = as_func_call)
}
\arguments{
\item{...}{The body of the rule, where character string with equal
to the name of a non-terminal will be treated as a non-terminal,
and other values are unaltered. In order to explicit mark a
character string as terminal, use the \code{term()} function.}

\item{name}{Optional character name to give to this rule. Note that
you should make sure different rules have distinct names. And
should avoid using the pattern of "non-terminal:number" as name,
because it is the pattern of auto-generated name in
\code{normalize_and_name_rules()}.}

\item{action}{A function that will construct the phenotype from a
list of constructed values of the rule body. The default is
\code{as_func_call} which will construct a function call where
the first element of the list will be treated as function using
\code{match.fun()}. Another useful action is \code{first} which
simply returns the first value in the list, and is useful when a
non-terminal is meant to give a leaf value instead of a function
call.}
}
\value{
An object of class "rule".
}
\description{
To construct one rule for a non-terminal.
}
