% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBGP.R
\name{simple_tournament}
\alias{simple_tournament}
\title{Simple tournament selection. Will randomly select two individuals,
and return the one with higher fitness. In case of ties, will
return the first selected one.}
\usage{
simple_tournament(pop, better_score = `>`, replace = TRUE)
}
\arguments{
\item{pop}{the list of current population of individuals, sorted
from the best to worst individuals, according to
\code{better_score}.}

\item{better_score}{Optional, defualt \code{>}. A function(fitness1,
fitness2) that should return TRUE if fitness1 is better than
fitness2. For example, \code{>} would maximize the fitness; while \code{<}
would minimize the fitness.}

\item{replace}{Optional, default TRUE. Whether the two individuals
are selected with replacement. Note that if TRUE, the two
selected individuals may be the same one, and therefore even the
individual with the worst fitness still has a chance of being
selected.}
}
\value{
An individual from pop.
}
\description{
Simple tournament selection. Will randomly select two individuals,
and return the one with higher fitness. In case of ties, will
return the first selected one.
}
