% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBGP.R
\name{sort_individuals}
\alias{sort_individuals}
\title{To sort a list of individuals from best to worst fitness.}
\usage{
sort_individuals(inds, maximize_fitness = TRUE)
}
\arguments{
\item{inds}{The list of individuals.}

\item{maximize_fitness}{Optional, default TRUE. If TRUE, to
maximize fitness; otherwise will minimize fitness.}
}
\value{
A sorted listed of individuals.
}
\description{
To sort a list of individuals from best to worst fitness.
}
