% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBGP.R
\name{stop_when_max_evals}
\alias{stop_when_max_evals}
\title{Stopping when reached maximum number of evaluations.}
\usage{
stop_when_max_evals(max_evals)
}
\arguments{
\item{max_evals}{The maximum number of evaluations.}
}
\value{
A stopping condition function that would return TRUE after
\code{max_evals} of evaluations have been performed.
}
\description{
Stopping when reached maximum number of evaluations.
}
