% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBGP.R
\name{stop_when_max_fitness}
\alias{stop_when_max_fitness}
\title{Stopping when reached max fitness value, in case we are maximizing
and we know what value is the optimum or good enough.}
\usage{
stop_when_max_fitness(max_fitness)
}
\arguments{
\item{max_fitness}{The fitness above which the evolution can stop.}
}
\value{
A stopping condition function that would return TRUE after
a fitness greater than or equal \code{max_fitness} is found.
}
\description{
Stopping when reached max fitness value, in case we are maximizing
and we know what value is the optimum or good enough.
}
