% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBGP.R
\name{stop_when_no_improvement_in_n_gen}
\alias{stop_when_no_improvement_in_n_gen}
\title{Stopping when no improvement in a certain number of generations}
\usage{
stop_when_no_improvement_in_n_gen(n_gen_no_improvement)
}
\arguments{
\item{n_gen_no_improvement}{To stop after this number of
generations without fitness improvement.}
}
\value{
A stopping condition function that would return TRUE if
there are no fitness improvement after
\code{n_gen_no_improvement} generations.
}
\description{
Stopping when no improvement in a certain number of generations
}
