% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBGP.R
\name{term}
\alias{term}
\title{To explicitly construct a terminal in a rule.}
\usage{
term(x)
}
\arguments{
\item{x}{The terminal value.}
}
\value{
An object of class "terminal", which is useful for
explicitly constructing terminals in a rule, in case it might be
regarded as a non-terminal.
}
\description{
To explicitly construct a terminal in a rule.
}
