% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBGP.R
\name{want_only_some_non_terminals}
\alias{want_only_some_non_terminals}
\title{A convenience function to generate node predicate that only wants
some non-terminals.}
\usage{
want_only_some_non_terminals(wanted_non_terminals)
}
\arguments{
\item{wanted_non_terminals}{The vector of non-terminals wanted.}
}
\value{
a predicate function that is TRUE if a node's non-terminal
is among the wanted list.
}
\description{
A convenience function to generate node predicate that only wants
some non-terminals.
}
