% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBGP.R
\name{always_true}
\alias{always_true}
\title{Function with constant output TRUE for an input.}
\usage{
always_true(...)
}
\arguments{
\item{...}{Place holder, dummy input(s).}
}
\value{
Constant TRUE.
}
\description{
Function with constant output TRUE for an input.
}
